version -events

+++ Core/Model.php
@@ -7,6 +7,21 @@ Class Model{
     public $list = [];
     public $primaryKey = "id";
     public $lastInsert = null;
+    
+    public function validate($data){
+        foreach($data as $k=>$v){ 
+            if(is_null($v)) return false;
+            $var = trim($v);
+            $var = ltrim($var);
+            if(!is_numeric($var) && empty($var)){
+                echo $var."<br>";
+                return false;
+            }
+        }
+
+        return true;
+    }
+
     function __construct($table, $pkey = "id"){
         //récupération des données
         $this->db = DataBase::getConnection();
@@ -28,11 +43,17 @@ Class Model{
         $this->list = $pre->fetchAll(PDO::FETCH_OBJ);
     }
 
-	function save($data){
+	function save($data, $inscription=false){
         // nom = elsa ; prenom = lorem 
-        echo "<pre>";
+        /*echo "<pre>";
         print_r($data);
-        echo "</pre>";
+        echo "</pre>";*/
+        $s = Session::loadSession(); 
+        if(!$inscription && !$s->loginVerification())return;
+        if(!$this->validate($data)){
+            $s->setFlash("Merci de vérifier que les champs sont bien rempli", "danger");
+            return false;
+        }
         
 		$key = $this->primaryKey;
 		$fields = array();
@@ -51,8 +72,8 @@ Class Model{
             $sql = "UPDATE ".$this->table." SET ".implode(',',$fields)." WHERE ".$key."=:".$key;
         }
 		if(isset($sql)){
-            print_r($sql);
-            print_r($w);
+            //print_r($sql);
+            //print_r($w);
 			$pre = $this->db->prepare($sql);
 			$pre->execute($w);
             if(!isset($data->$key)){
@@ -65,14 +86,21 @@ Class Model{
 	}
 	
 	function delete($id){
+        $s = Session::loadSession();
+        if(!$s->loginVerification())return;
 		$sql = "DELETE FROM {$this->table} WHERE {$this->primaryKey} = $id";
 		$this->db->query($sql);
 	}
 
     public function exec($sql, $type = true){
+        $select = explode(" ", $sql);
+        if(count($select) === 0 || strtolower($select[0])!=="select"){
+            $s = Session::loadSession();
+            if(!$s->loginVerification())return;
+        }
         if(isset($sql)){
 			$pre = $this->db->prepare($sql);
-			$pre->execute($w);
+			$pre->execute();
             if($type) $this->list = $pre->fetchAll(PDO::FETCH_OBJ);
             else $this->lastInsert = $this->db->lastInsertId();
 			return true;
